fix a bug when the SelectedMovieInfo component is mounted and then immediatelt unmounted and again immediately mounted (react strict mode)
two browser APIs try to fetch the data. The first one receives the signal from the clean up
callback function, so it cancels the fetching process. Therefore an exception is thrown and
the catch block and then the finally block are executed which sets the state isLoading to false.
The second browser API started the execution of the async function at the same time with the
first API and set the state isLoading to true. If the first API sets the state to false after the first API set it to true, then the UI will be painted based on the value false
which is wrong.

to fix, either setIsLoading(false) after a successful fect and also after a failed fetch only if
(error.name !=='AbortError') or use the flag isMounted which is set to true at the beginning of
the useEffect callback function and is set to false in the clean up callback function. The states
must be updated only if the component isMounted.